+++ tap_xero/client.py
@@ -236,7 +236,7 @@ class XeroClient():
             raise_for_error(response)
 
     @backoff.on_exception(backoff.expo, XeroUnauthorizedError, max_tries=1)
-    @backoff.on_exception(backoff.expo, (json.decoder.JSONDecodeError, XeroInternalError, XeroUnauthorizedError), max_tries=3)
+    @backoff.on_exception(backoff.expo, (json.decoder.JSONDecodeError, XeroInternalError), max_tries=3)
     @backoff.on_exception(retry_after_wait_gen, XeroTooManyInMinuteError, giveup=is_not_status_code_fn([429]), jitter=None, max_tries=3)
     def filter(self, tap_stream_id, since=None, **params):
         xero_resource_name = tap_stream_id.title().replace("_", "")